@@ -57,10 +57,11 @@
     "@plus-pro-components/types": "workspace:*",
     "@plus-pro-components/utils": "workspace:*",
     "dayjs": "^1.11.7",
-    "element-plus": "^2.5.3",
+    "element-plus": "^2.6.3",
     "lodash-es": "^4.17.21",
     "sortablejs": "^1.15.0",
-    "vue": "^3.3.4",
+    "vite": "^5.2.9",
+    "vue": "^3.4.21",
     "vue-tsc": "1.8.27"
   },
   "devDependencies": {
@@ -69,17 +70,17 @@
     "@inquirer/prompts": "^1.2.3",
     "@pnpm/find-workspace-packages": "^6.0.8",
     "@pnpm/types": "^9.0.0",
-    "@rollup/plugin-commonjs": "^25.0.0",
-    "@rollup/plugin-node-resolve": "^15.0.2",
+    "@rollup/plugin-commonjs": "^25.0.7",
+    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/fs-extra": "^11.0.1",
     "@types/jsdom": "^21.1.1",
     "@types/lodash-es": "^4.17.7",
     "@types/node": "^20.1.4",
     "@types/sortablejs": "^1.15.1",
-    "@vitejs/plugin-vue": "^4.2.3",
-    "@vitejs/plugin-vue-jsx": "^3.0.1",
-    "@vitest/coverage-c8": "^0.31.1",
-    "@vue/test-utils": "^2.3.2",
+    "@vitejs/plugin-vue": "^5.0.4",
+    "@vitejs/plugin-vue-jsx": "^3.1.0",
+    "@vitest/coverage-v8": "^1.5.0",
+    "@vue/test-utils": "^2.4.5",
     "autoprefixer": "^10.4.14",
     "chalk": "4.1.2",
     "commitizen": "^4.3.0",
@@ -103,26 +104,24 @@
     "postcss": "^8.4.24",
     "prettier": "^2.8.8",
     "rimraf": "^5.0.0",
-    "rollup": "^3.22.0",
-    "rollup-plugin-banner2": "^1.2.2",
-    "rollup-plugin-esbuild": "^5.0.0",
+    "rollup": "^4.14.3",
+    "rollup-plugin-banner2": "^1.2.3",
+    "rollup-plugin-esbuild": "^6.1.1",
     "rollup-plugin-postcss": "^4.0.2",
     "rollup-plugin-scss": "^4.0.0",
-    "rollup-plugin-vue": "^6.0.0",
-    "sass": "^1.69.4",
+    "sass": "^1.75.0",
     "semver": "^7.5.1",
-    "ts-morph": "^18.0.0",
-    "ts-node": "^10.9.1",
-    "tsx": "^3.14.0",
-    "typescript": "^5.1.6",
+    "tsx": "^4.7.2",
+    "typescript": "^5.4.5",
     "unbuild": "^2.0.0",
-    "vitest": "^0.31.1"
+    "vite-plugin-dts": "^3.8.3",
+    "vitest": "^1.5.0"
   },
   "peerDependencies": {
     "vue": "^3.2.0"
   },
   "engines": {
-    "node": ">= 16"
+    "node": ">= 18"
   },
   "config": {
     "commitizen": {
