@@ -16,7 +16,7 @@ describe('page/index.vue', () => {
     ) => {
       const { page = 1, pageSize = 20, status, name } = query || {}
       const total = 1000
-      const List = [...new Array(total)].map((item, index) => {
+      const List = Array.from({ length: total }).map((item, index) => {
         return {
           id: index,
           name: index === 0 ? 'name'.repeat(20) : index + 'name',
@@ -190,7 +190,7 @@ describe('page/index.vue', () => {
     ) => {
       const { page = 1, pageSize = 20, status, name } = query || {}
       const total = 1000
-      const List = [...new Array(total)].map((item, index) => {
+      const List = Array.from({ length: total }).map((item, index) => {
         return {
           id: index,
           name: index === 0 ? 'name'.repeat(20) : index + 'name',
